test: update access token
* Regenerated access key and token.

* Moved oauth2 tests into its own file.

* Refactored common functions into `test/backend/utils.js`.

* Added script to run oauth2 tests only.
