+++ test/backend/utils.js
@@ -1,11 +1,74 @@
 const fs = require('fs');
 
+const fetchCookie = require('fetch-cookie');
+
+const nodeFetch = require('node-fetch');
+let fetch = fetchCookie(nodeFetch);
+
 const restify = require('restify');
 
 import {dissoc, merge} from 'ramda';
 
 import {PlotlyAPIRequest} from '../../backend/persistent/PlotlyAPI.js';
 import {getSetting} from '../../backend/settings.js';
+
+// Helper functions
+export function clearCookies() {
+    fetch = fetchCookie(nodeFetch);
+}
+
+export function GET(path) {
+    return fetch(`http://localhost:9494/${path}`, {
+        method: 'GET',
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        }
+    });
+}
+
+export function PATCH(path, body = {}) {
+    return fetch(`http://localhost:9494/${path}`, {
+        method: 'PATCH',
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        },
+        body: body ? JSON.stringify(body) : null
+    });
+}
+
+export function POST(path, body = {}) {
+    return fetch(`http://localhost:9494/${path}`, {
+        method: 'POST',
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        },
+        body: body ? JSON.stringify(body) : null
+    });
+}
+
+export function PUT(path, body = {}) {
+    return fetch(`http://localhost:9494/${path}`, {
+        method: 'PUT',
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        },
+        body: body ? JSON.stringify(body) : null
+    });
+}
+
+export function DELETE(path) {
+    return fetch(`http://localhost:9494/${path}`, {
+        method: 'DELETE',
+        headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+        }
+    });
+}
 
 // Returns a function to be used in a Promise chain that asserts a response status
 export function assertResponseStatus(expectedStatus) {
@@ -72,17 +135,22 @@ export const columns = [
 
 // test account on prod
 export const username = 'plotly-database-connector';
-export const apiKey = 'reiptow6gu';
+export const apiKey = 'ccJMY9txxWhRRMWNLgOT';
 /*
  * This is a real live access token associated with
  * the user account plotly-database-connector on
  * https://plot.ly.
  *
- * This token is generated by visiting
- * https://plot.ly/o/authorize/?response_type=token&client_id=isFcew9naom2f1khSiMeAtzuOvHXHuLwhPsM7oPt&redirect_uri=http://localhost:9494/oauth2/callback
- * in your web browser.
+ * Follow the below steps, if you ever need to regenerate the access token:
+ * - rm -rf ~/.plotly/connector
+ * - yarn run start
+ * - Add a connector (e.g. DB2)
+ * - Click on the tab Plot.ly
+ * - Click on Login into Plot.ly
+ * - Login as username 'plotly-database-connector`
+ * - Once authorized, the access token can be found in the file ~/.plotly/connector/settings.yaml
  */
-export const accessToken = '2MiYq1Oive6RRjC6y9D4u7DjlXSs7k';
+export const accessToken = 'UcTdbxMlXyIWY45GOxyRALQhrXVIeq';
 
 // This fid with these UIDs actually exist on prod
 export const validFid = 'plotly-database-connector:197';
